requirements.txt
numpy>=1.24
torch>=2.0
torchvision>=0.15
opencv-python>=4.8
imageio>=2.31
imageio-ffmpeg>=0.4.9
pyyaml>=6.0
tqdm>=4.66
